refactor(hw/cpu): split files
